feat: standardize gallery cards into a shared component
