+++ frontend/src/styles/components/AstroGallery.module.css
@@ -27,31 +27,6 @@
   justify-content: center;
 }
 
-.gridItem {
-  width: 100%;
-  aspect-ratio: 1 / 1;
-  /* Maintain square shape */
-  background-color: var(--portfolio-bg-light);
-  border-radius: var(--radius-lg);
-  overflow: hidden;
-  box-shadow: var(--shadow-md);
-  transition: transform var(--transition-normal);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.gridItem:hover {
-  transform: scale(1.03);
-}
-
-.gridItem img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  display: block;
-}
-
 .placeholder {
   font-size: 1.2rem;
 }
@@ -129,109 +104,6 @@
   border-color: var(--accent-gold) !important;
   font-weight: var(--font-weight-bold);
   box-shadow: 0 4px 16px rgba(255, 184, 0, 0.12);
-}
-
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background: rgba(0, 0, 0, 0.85);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: var(--z-modal);
-  overflow-y: auto;
-}
-
-.modalContent {
-  background: var(--portfolio-bg-dark);
-  border-radius: var(--radius-xl);
-  padding: var(--space-8) var(--space-8) var(--space-6) var(--space-8);
-  box-shadow: var(--shadow-xl);
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-width: 320px;
-  max-width: 1000px;
-  width: 90vw;
-  margin: 5vh 0;
-  /* Add vertical margin for top and bottom space */
-  max-height: 90vh;
-  /* Prevent modal from exceeding viewport */
-  overflow-y: auto;
-  /* Make modal scrollable if content is too tall */
-  -ms-overflow-style: none;
-  /* IE and Edge */
-  scrollbar-width: none;
-  /* Firefox */
-}
-
-.modalContent::-webkit-scrollbar {
-  display: none;
-}
-
-.modalImage {
-  max-width: 90%;
-  border-radius: 8px;
-  margin-bottom: 0.5rem;
-  object-fit: contain;
-  display: block;
-  color: #94a3b8;
-  font-size: 0.875rem;
-}
-
-.tagsContainer {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-  margin-left: 1rem;
-}
-
-.tagBadge {
-  background: rgba(56, 189, 248, 0.1);
-  color: var(--celestial-blue);
-  padding: 0.25rem 0.75rem;
-  border-radius: 9999px;
-  font-size: 0.75rem;
-  font-weight: 500;
-  border: 1px solid rgba(56, 189, 248, 0.2);
-}
-
-.modalDescription {
-  color: #fff;
-  font-size: 1.1rem;
-  text-align: center;
-  margin-top: 0.5rem;
-  font-weight: 400;
-  font-family: "Roboto", Arial, Helvetica, sans-serif;
-  padding: 1.5rem 2rem;
-  background: rgba(24, 28, 43, 0.85);
-  border-radius: 8px;
-  box-sizing: border-box;
-  word-break: break-word;
-  white-space: pre-line;
-  max-width: 100%;
-  overflow-wrap: break-word;
-}
-
-.modalClose {
-  position: absolute;
-  top: 0.5rem;
-  right: 1rem;
-  background: none;
-  border: none;
-  color: #fff;
-  font-size: 2rem;
-  cursor: pointer;
-  z-index: 10;
-  transition: color 0.2s;
-}
-
-.modalClose:hover {
-  color: var(--accent-gold);
 }
 
 .noResults {
