+++ frontend/src/styles/components/GalleryCard.module.css
@@ -1,0 +1,153 @@
+.card {
+  position: relative;
+  height: 400px;
+  width: 100%;
+  border-radius: 1.5rem;
+  overflow: hidden;
+  background: rgba(15, 23, 42, 0.6);
+  backdrop-filter: blur(12px);
+  -webkit-backdrop-filter: blur(12px);
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  cursor: pointer;
+  transition: all 0.7s var(--transition-normal);
+  text-align: left;
+  padding: 0;
+}
+
+.imageWrapper {
+  position: absolute;
+  inset: 0;
+  background: #02040a;
+  overflow: hidden;
+}
+
+.placeholder {
+  position: absolute;
+  inset: 0;
+  background: rgba(15, 23, 42, 0.4);
+  backdrop-filter: blur(20px);
+  -webkit-backdrop-filter: blur(20px);
+  transition: opacity 1s ease-out;
+  z-index: 1;
+}
+
+.placeholder.hide {
+  opacity: 0;
+}
+
+.cardImage {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  opacity: 0;
+  transition:
+    opacity 1s ease-out,
+    transform 0.7s var(--transition-normal);
+}
+
+.cardImage.show {
+  opacity: 0.6;
+}
+
+.card:hover .cardImage.show {
+  opacity: 0.8;
+  transform: scale(1.05);
+}
+
+.newBadge {
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+  background: var(--celestial-blue);
+  color: #02040a;
+  padding: 0.25rem 0.75rem;
+  border-radius: 9999px;
+  font-size: 0.625rem;
+  font-weight: 800;
+  text-transform: uppercase;
+  letter-spacing: 0.1em;
+  z-index: 10;
+  box-shadow: 0 0 15px rgba(56, 189, 248, 0.4);
+  font-family: var(--font-family-heading);
+}
+
+.cardContent {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  padding: 2.5rem;
+  transform: translateY(2rem);
+  transition: transform 0.6s var(--transition-normal);
+  z-index: 5;
+  box-sizing: border-box;
+}
+
+.card:hover .cardContent {
+  transform: translateY(0);
+}
+
+.category {
+  font-size: 0.5625rem;
+  /* 9px */
+  text-transform: uppercase;
+  letter-spacing: 0.4em;
+  color: var(--celestial-blue);
+  font-weight: 700;
+  display: block;
+  margin-bottom: 0.75rem;
+  font-family: var(--font-family-heading);
+}
+
+.cardTitle {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: var(--celestial-white);
+  line-height: 1.2;
+  font-family: var(--font-family-heading);
+  margin-bottom: 0.5rem;
+}
+
+.cardLocation {
+  display: flex;
+  align-items: center;
+  gap: 0.4rem;
+  font-size: 0.75rem;
+  color: var(--celestial-white);
+  opacity: 0.8;
+  font-family: var(--font-family-mono);
+  margin-bottom: 0.75rem;
+}
+
+.metaIcon {
+  opacity: 0.6;
+}
+
+.cardDescription {
+  font-size: 0.875rem;
+  color: var(--celestial-white);
+  opacity: 0;
+  max-height: 0;
+  overflow: hidden;
+  line-height: 1.5;
+  transition: all 0.5s var(--transition-normal);
+  font-family: var(--font-family-primary);
+}
+
+.card:hover .cardDescription {
+  opacity: 0.7;
+  max-height: 100px;
+  margin-top: 0.5rem;
+}
+
+.divider {
+  margin-top: 1.5rem;
+  height: 1px;
+  width: 0;
+  background: rgba(255, 255, 255, 0.2);
+  transition: width 0.7s var(--transition-normal);
+}
+
+.card:hover .divider {
+  width: 100%;
+}
